Module tower_web_protobuf::common

Structs

FutErr
A future representing a value that is immediately ready.

FutOk
A future representing a value that is immediately ready.

FutureResult
A future representing a value that is immediately ready.

HeaderName
Represents an HTTP header field name

HttpRequest
Represents an HTTP request.

HttpResponse
Represents an HTTP response

ResponseFuture

Traits

Future
Trait for types which are a placeholder of a value that may become available at some later point in time.

Service
An asynchronous function from Request to a Response.

Type Definitions

Poll
Return type of the Future::poll method, indicates whether a future's value is ready or not.