Struct tower_web_protobuf::types::proto_types::Proto

pub struct Proto<M: MessagePlus>(pub M);

A wrapper struct for a protobuf message type.

This has to exist because impl<T> Trait for T requires T to be 'covered' by a local type (i.e. Proto), when Self is used (I think). Self is used by Extract (Future: ExtractFuture<Item = Self>) which I think is why:

impl<M, B: BufStream> Extract<B> for M
where
    M: 'static + Message + MessageWrapper<M>
{
    type Future = Immediate<M>;
}

doesn't work.

Niko's excellent blog post has a full writeup.

The effect of this is that you'll have to specify Proto instead of just Message in your functions within impl_web!(). This, in turn hurts testability (you have to wrap your Messages to pass them into the endpoint functions) and kind of works counter to the PORTs (plain old Rust types) philosophy. But, I'm pretty sure it's the best we can do without using macros or modifying tower-web.

Into for Proto and From for Proto are implemented to ease the pain a little bit. Into for Proto instead of From<Proto> for M for the same reasons we aren't just implementing Extract for M.

Deref is also implemented it should be possible to use Proto as an M for most everything.

Methods

impl<M: MessagePlus> Proto<M>

pub fn move_inner(self) -> M

Unwraps the message (of type M) from it's Proto container, consuming the container in the process.

pub fn new(message: M) -> Self

Creates a new Proto instance from a message of type M.